+++ R/validate_inputs.R
@@ -91,8 +91,8 @@ do_validate_population <- function(population) {
     colnames(population),
     c("area_id", "calendar_quarter", "sex", "age_group", "source", "population"))
   assert_calendar_quarter_column(population)
-  assert_expected_values(population, "sex", c("male", "female"), all_values=TRUE)
-  assert_expected_values(population, "age_group", naomi::get_five_year_age_groups(), all_values=TRUE)
+  assert_expected_values(population, "sex", c("male", "female"), all_values = TRUE)
+  assert_expected_values(population, "age_group", naomi::get_five_year_age_groups(), all_values = TRUE)
   assert_unique_combinations(population, c("area_id", "calendar_quarter", "sex", "age_group"))
   assert_single_source(population)
   assert_no_na(population, "population")
@@ -110,9 +110,6 @@ do_validate_population <- function(population) {
 #'
 #' @param shape A file object (path, hash, filename) corresponding to
 #'   the input shape file.
-#'
-#' @param pjnz A file object (path, hash, filename) corresponding to
-#'   the input PJNZ file.
 #'
 #' @param strict If FALSE then run less stringent validation rules, used
 #'   for data exploration mode.
@@ -154,9 +151,6 @@ do_validate_programme <- function(programme, shape, strict = TRUE) {
 #'
 #' @param shape A file object (path, hash, filename) corresponding to
 #'   the input shape file.
-#'
-#' @param pjnz A file object (path, hash, filename) corresponding to
-#'   the input PJNZ file.
 #'
 #' @param strict If FALSE then run less stringent validation rules, used
 #'   for data exploration mode.
@@ -202,9 +196,6 @@ do_validate_anc <- function(anc, shape, strict = TRUE) {
 #' @param strict If FALSE then run less stringent validation rules, used
 #'   for data exploration mode.
 #'
-#' @param pjnz A file object (path, hash, filename) corresponding to
-#'   the input PJNZ file.
-#'
 #' @return An error if invalid.
 #' @keywords internal
 do_validate_survey <- function(survey, shape, strict = TRUE) {
@@ -232,6 +223,40 @@ do_validate_survey <- function(survey, shape, strict = TRUE) {
        filters = list("age" = get_age_filters(data),
                       "surveys" = get_survey_filters(data),
                       "indicators" = get_indicator_filters(data, "survey")),
+       warnings = list())
+}
+
+
+#' Validate VMMC data file.
+#'
+#' Check that VMMC data file can be read.
+#'
+#' @param vmmc A file object (path, hash, filename) corresponding to
+#'   the input VMMC file.
+#'
+#' @param shape A file object (path, hash, filename) corresponding to
+#'   the input shape file.
+#'
+#' @param strict If FALSE then run less stringent validation rules, used
+#'   for data exploration mode.
+#'
+#' @return An error if invalid.
+#' @keywords internal
+do_validate_vmmc <- function(vmmc, shape, strict = TRUE) {
+  assert_file_extension(vmmc, "xlsx")
+  assert_sheet_exists(vmmc$path, "Datapack inputs")
+  ## Skip the first header, the file has two rows of headers
+  data <- readxl::read_xlsx(vmmc$path, sheet = "Datapack inputs", skip = 1,
+                            .name_repair = "minimal")
+  assert_single_country(data, "vmmc")
+  assert_column_names(colnames(data), c("area_id", "15-24", "25-34",
+                      "35-49", "50+", "15-24", "25-34", "35-49", "50+"))
+  shape_regions <- read_regions(shape, "shape")
+  assert_consistent_regions(shape_regions$area_id,
+                            unique(data$area_id),
+                            "vmmc")
+  list(data = json_verbatim("null"),
+       filters = json_verbatim("null"),
        warnings = list())
 }