+++ tests/testthat/test-model-options.R
@@ -5,10 +5,12 @@ test_that("do_endpoint_model_options correctly builds options and fallbacks", {
   anc <- file_object(test_path("testdata", "anc.csv"))
 
   mock_get_controls_json <- mockery::mock('"{"test"}')
-  with_mock(get_controls_json = mock_get_controls_json,  {
-    json <- do_endpoint_model_options(shape, survey, art, anc)
-    args <- mockery::mock_args(mock_get_controls_json)
-  })
+  with_mocked_bindings(
+    {
+      json <- do_endpoint_model_options(shape, survey, art, anc)
+      args <- mockery::mock_args(mock_get_controls_json)
+    },
+    get_controls_json = mock_get_controls_json)
   options <- args[[1]][[3]]
   expect_setequal(names(options),
                   c("area_scope", "area_level",
@@ -77,16 +79,13 @@ test_that("do_endpoint_model_options correctly builds options and fallbacks", {
                   c("area_scope",
                     "calendar_quarter_t1", "survey_prevalence",
                     "survey_art_coverage", "anc_prevalence_year1",
-                    "anc_prevalence_year2", "anc_art_coverage_year1",
-                    "anc_art_coverage_year2"))
+                    "anc_art_coverage_year1"))
   expect_equal(overrides$area_scope, scalar("MWI"))
   expect_equal(overrides$calendar_quarter_t1, scalar("CY2016Q1"))
   expect_equal(overrides$survey_prevalence, scalar("DEMO2016PHIA"))
   expect_equal(overrides$survey_art_coverage, scalar("DEMO2016PHIA"))
   expect_equal(overrides$anc_prevalence_year1, scalar("2016"))
-  expect_equal(overrides$anc_prevalence_year2, scalar(""))
   expect_equal(overrides$anc_art_coverage_year1, scalar("2016"))
-  expect_equal(overrides$anc_art_coverage_year2, scalar(""))
 })
 
 test_that("do_endpoint_model_options without programme data", {
@@ -94,10 +93,12 @@ test_that("do_endpoint_model_options without programme data", {
   survey <- file_object(file.path("testdata", "survey.csv"))
 
   mock_get_controls_json <- mockery::mock('"{"test"}')
-  with_mock(get_controls_json = mock_get_controls_json,  {
-    json <- do_endpoint_model_options(shape, survey, NULL, NULL)
-    args <- mockery::mock_args(mock_get_controls_json)
-  })
+  with_mocked_bindings(
+    {
+      json <- do_endpoint_model_options(shape, survey, NULL, NULL)
+      args <- mockery::mock_args(mock_get_controls_json)
+    },
+    get_controls_json = mock_get_controls_json)
   options <- args[[1]][[3]]
   expect_setequal(names(options),
                   c("area_scope", "area_level",
@@ -156,21 +157,21 @@ test_that("do_endpoint_model_options without programme data", {
 })
 
 test_that("do_endpoint_model_options overrides anc year2 to 2022 if in data", {
-  shape <- file_object(file.path("testdata", "malawi.geojson"))
-  survey <- file_object(file.path("testdata", "survey.csv"))
-  art <- file_object(file.path("testdata", "programme.csv"))
-  anc <- file_object(file.path("testdata", "anc.csv"))
+  shape <- file_object(test_path("testdata", "malawi.geojson"))
+  survey <- file_object(test_path("testdata", "survey.csv"))
+  art <- file_object(test_path("testdata", "programme.csv"))
+  anc <- file_object(test_path("testdata", "anc.csv"))
 
   mock_get_controls_json <- mockery::mock('"{"test"}')
   mock_get_years <- mockery::mock(c(2022, 2021, 2020, 2019))
-  with_mock(get_controls_json = mock_get_controls_json,
-            get_years = mock_get_years, {
-    json <- do_endpoint_model_options(shape, survey, art, anc)
-    args <- mockery::mock_args(mock_get_controls_json)
-  })
+  with_mocked_bindings(
+    {
+      json <- do_endpoint_model_options(shape, survey, art, anc)
+      args <- mockery::mock_args(mock_get_controls_json)
+    },
+    get_controls_json = mock_get_controls_json,
+    get_years = mock_get_years)
   overrides <- args[[1]][[4]]
-  expect_equal(overrides$anc_prevalence_year2, scalar("2022"))
-  expect_equal(overrides$anc_art_coverage_year2, scalar("2022"))
   ## Year 1 defaults are NULL as survey year not in ANC years
   expect_equal(overrides$anc_prevalence_year1, scalar(""))
   expect_equal(overrides$anc_art_coverage_year1, scalar(""))
@@ -352,10 +353,12 @@ test_that("model options work when survey_mid_calendar_quarter missing", {
   survey$path <- t
 
   mock_get_controls_json <- mockery::mock('"{"test"}')
-  with_mock(get_controls_json = mock_get_controls_json,  {
-    json <- do_endpoint_model_options(shape, survey, art, anc)
-    args <- mockery::mock_args(mock_get_controls_json)
-  })
+  with_mocked_bindings(
+    {
+      json <- do_endpoint_model_options(shape, survey, art, anc)
+      args <- mockery::mock_args(mock_get_controls_json)
+    },
+    get_controls_json = mock_get_controls_json)
 
   ## Fallback set to most recent time option
   fallback <- args[[1]][[4]]
@@ -410,7 +413,7 @@ test_that("getting survey options for missing indicator returns empty values", {
   expect_equal(
     get_survey_options(data, metadata, "art_coverage"),
     list(
-      options = NULL,
+      options = list(),
       default = scalar("")
     )
   )